Start migration to TvDB API v2
