+++ Shoko.Server/Providers/TvDB/TvDBHelper.cs
@@ -146,59 +146,6 @@ namespace Shoko.Models.TvDB
             }
         }
 
-/*
-		public static bool ConfirmTvDBOnline()
-		{
-			TvDB_Series tvser = GetSeriesInfoOnline(73255);
-			if (tvser == null)
-				return false;
-			else
-				return true;
-		}
-        */
-
-        public static TvDB_Series GetSeriesInfoOnline(int seriesID)
-        {
-            try
-            {
-                //Init();
-
-                string url = string.Format(Shoko.Server.Constants.TvDBURLs.urlSeriesBaseXML, URLMirror,
-                    Shoko.Server.Constants.TvDBURLs.apiKey,
-                    seriesID,
-                    ServerSettings.TvDB_Language);
-                logger.Trace("GetSeriesInfo: {0}", url);
-
-                // Search for a series
-                string xmlSeries = Utils.DownloadWebPage(url);
-                logger.Trace("GetSeriesInfo RESULT: {0}", xmlSeries);
-
-                if (xmlSeries.Trim().Length == 0) return null;
-
-                XmlDocument docSeries = new XmlDocument();
-                docSeries.LoadXml(xmlSeries);
-
-                TvDB_Series tvSeries = null;
-                if (docSeries != null)
-                {
-                    tvSeries = RepoFactory.TvDB_Series.GetByTvDBID(seriesID);
-                    if (tvSeries == null)
-                        tvSeries = new TvDB_Series();
-
-                    tvSeries.PopulateFromSeriesInfo(docSeries);
-                    RepoFactory.TvDB_Series.Save(tvSeries);
-                }
-
-                return tvSeries;
-            }
-            catch (Exception ex)
-            {
-                logger.Error(ex, "Error in TVDBHelper.GetSeriesInfoOnline: " + ex.ToString());
-            }
-
-            return null;
-        }
-
         public XmlDocument GetSeriesBannersOnline(int seriesID)
         {
             try
@@ -663,49 +610,6 @@ namespace Shoko.Models.TvDB
             return banners;
         }
 
-
-        public List<TVDB_Series_Search_Response> SearchSeries(string criteria)
-        {
-            List<TVDB_Series_Search_Response> results = new List<TVDB_Series_Search_Response>();
-
-            try
-            {
-                Init();
-
-                if (!initialised) return results;
-
-                // Search for a series
-                string url = string.Format(Shoko.Server.Constants.TvDBURLs.urlSeriesSearch, criteria);
-                logger.Trace("Search TvDB Series: {0}", url);
-
-                string xmlSeries = Utils.DownloadWebPage(url);
-
-                XmlDocument docSeries = new XmlDocument();
-                docSeries.LoadXml(xmlSeries);
-
-                bool hasData = docSeries["Data"].HasChildNodes;
-                if (hasData)
-                {
-                    XmlNodeList seriesItems = docSeries["Data"].GetElementsByTagName("Series");
-
-                    foreach (XmlNode series in seriesItems)
-                    {
-                        TVDB_Series_Search_Response searchResult = new TVDB_Series_Search_Response();
-                        searchResult.Populate(series);
-
-                        results.Add(searchResult);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                logger.Error(ex, "Error in SearchSeries: " + ex.ToString());
-            }
-
-            return results;
-        }
-
-
         public static List<int> GetUpdatedSeriesList(string serverTime)
         {
             List<int> seriesList = new List<int>();
@@ -840,93 +744,6 @@ namespace Shoko.Models.TvDB
             }
         }
 
-
-        public static string LinkAniDBTvDB(int animeID, enEpisodeType aniEpType, int aniEpNumber, int tvDBID,
-            int tvSeasonNumber, int tvEpNumber, bool excludeFromWebCache, bool additiveLink = false)
-        {
-            using (var session = DatabaseFactory.SessionFactory.OpenSession())
-            {
-                if (!additiveLink)
-                    // remove all current links
-                    RemoveAllAniDBTvDBLinks(session.Wrap(), animeID);
-
-                // check if we have this information locally
-                // if not download it now
-                TvDB_Series tvSeries = RepoFactory.TvDB_Series.GetByTvDBID(tvDBID);
-                if (tvSeries == null)
-                {
-                    // we download the series info here just so that we have the basic info in the
-                    // database before the queued task runs later
-                    tvSeries = GetSeriesInfoOnline(tvDBID);
-                }
-
-                // download and update series info, episode info and episode images
-                // will also download fanart, posters and wide banners
-                CommandRequest_TvDBUpdateSeriesAndEpisodes cmdSeriesEps =
-                    new CommandRequest_TvDBUpdateSeriesAndEpisodes(tvDBID,
-                        false);
-                //Optimize for batch updates, if there are a lot of LinkAniDBTvDB commands queued 
-                //this will cause only one updateSeriesAndEpisodes command to be created
-                if (RepoFactory.CommandRequest.GetByCommandID(cmdSeriesEps.CommandID) == null)
-                {
-                    cmdSeriesEps.Save();
-                }
-
-                CrossRef_AniDB_TvDBV2 xref = RepoFactory.CrossRef_AniDB_TvDBV2.GetByTvDBID(session, tvDBID,
-                    tvSeasonNumber, tvEpNumber,
-                    animeID,
-                    (int) aniEpType, aniEpNumber);
-                if (xref == null)
-                    xref = new CrossRef_AniDB_TvDBV2();
-
-                xref.AnimeID = animeID;
-                xref.AniDBStartEpisodeType = (int) aniEpType;
-                xref.AniDBStartEpisodeNumber = aniEpNumber;
-
-                xref.TvDBID = tvDBID;
-                xref.TvDBSeasonNumber = tvSeasonNumber;
-                xref.TvDBStartEpisodeNumber = tvEpNumber;
-                if (tvSeries != null)
-                    xref.TvDBTitle = tvSeries.SeriesName;
-
-                if (excludeFromWebCache)
-                    xref.CrossRefSource = (int) CrossRefSource.WebCache;
-                else
-                    xref.CrossRefSource = (int) CrossRefSource.User;
-
-                RepoFactory.CrossRef_AniDB_TvDBV2.Save(xref);
-
-                SVR_AniDB_Anime.UpdateStatsByAnimeID(animeID);
-
-                logger.Trace("Changed tvdb association: {0}", animeID);
-
-                if (!excludeFromWebCache)
-                {
-                    var req = new CommandRequest_WebCacheSendXRefAniDBTvDB(xref.CrossRef_AniDB_TvDBV2ID);
-                    req.Save();
-                }
-
-                if (ServerSettings.Trakt_IsEnabled && !string.IsNullOrEmpty(ServerSettings.Trakt_AuthToken))
-                {
-                    // check for Trakt associations
-                    List<CrossRef_AniDB_TraktV2> trakt = RepoFactory.CrossRef_AniDB_TraktV2.GetByAnimeID(animeID);
-                    if (trakt.Count != 0)
-                    {
-                        // remove them and rescan
-                        foreach (CrossRef_AniDB_TraktV2 a in trakt)
-                        {
-                            RepoFactory.CrossRef_AniDB_TraktV2.Delete(a);
-                        }
-                    }
-
-                    var cmd2 = new CommandRequest_TraktSearchAnime(animeID, false);
-                    cmd2.Save(session);
-                }
-            }
-
-            return "";
-        }
-
         public static void LinkAniDBTvDBEpisode(int aniDBID, int tvDBID, int animeID)
         {
             CrossRef_AniDB_TvDB_Episode xref = RepoFactory.CrossRef_AniDB_TvDB_Episode.GetByAniDBEpisodeID(aniDBID);
@@ -966,76 +783,7 @@ namespace Shoko.Models.TvDB
             req.Save();
         }
 
-        public static void RemoveAllAniDBTvDBLinks(ISessionWrapper session, int animeID, int aniEpType = -1)
-        {
-            // check for Trakt associations
-            List<CrossRef_AniDB_TraktV2> trakt = RepoFactory.CrossRef_AniDB_TraktV2.GetByAnimeID(animeID);
-            if (trakt.Count != 0)
-            {
-                // remove them and rescan
-                foreach (CrossRef_AniDB_TraktV2 a in trakt)
-                {
-                    RepoFactory.CrossRef_AniDB_TraktV2.Delete(a);
-                }
-            }
-
-            List<CrossRef_AniDB_TvDBV2> xrefs = RepoFactory.CrossRef_AniDB_TvDBV2.GetByAnimeID(session, animeID);
-            if (xrefs == null || xrefs.Count == 0) return;
-
-            foreach (CrossRef_AniDB_TvDBV2 xref in xrefs)
-            {
-                if (aniEpType != -1 && aniEpType == xref.AniDBStartEpisodeType) continue;
-
-                RepoFactory.CrossRef_AniDB_TvDBV2.Delete(xref.CrossRef_AniDB_TvDBV2ID);
-
-                if (aniEpType == -1)
-                {
-                    foreach (enEpisodeType eptype in Enum.GetValues(typeof(enEpisodeType)))
-                    {
-                        CommandRequest_WebCacheDeleteXRefAniDBTvDB req = new CommandRequest_WebCacheDeleteXRefAniDBTvDB(
-                            animeID,
-                            (int) eptype, xref.AniDBStartEpisodeNumber,
-                            xref.TvDBID, xref.TvDBSeasonNumber, xref.TvDBStartEpisodeNumber);
-                        req.Save();
-                    }
-                }
-                else
-                {
-                    CommandRequest_WebCacheDeleteXRefAniDBTvDB req = new CommandRequest_WebCacheDeleteXRefAniDBTvDB(
-                        animeID,
-                        aniEpType, xref.AniDBStartEpisodeNumber,
-                        xref.TvDBID, xref.TvDBSeasonNumber, xref.TvDBStartEpisodeNumber);
-                    req.Save();
-                }
-            }
-
-            SVR_AniDB_Anime.UpdateStatsByAnimeID(animeID);
-        }
-
-        /*
-        public static void DownloadAllEpisodes()
-        {
-            CrossRef_AniDB_TvDBV2Repository repCrossRef = new CrossRef_AniDB_TvDBV2Repository();
-            List<CrossRef_AniDB_TvDBV2> allCrossRefs = repCrossRef.GetAll();
-
-            List<int> tvDBIDs = new List<int>();
-            foreach (CrossRef_AniDB_TvDBV2 xref in allCrossRefs)
-            {
-                if (!tvDBIDs.Contains(xref.TvDBID)) tvDBIDs.Add(xref.TvDBID);
-            }
-
-            DownloadAllEpisodes(tvDBIDs);
-        }
-
-        public static void DownloadAllEpisodes(List<int> tvDBIDs)
-        {
-            foreach (int tvid in tvDBIDs)
-            {
-                CommandRequest_TvDBUpdateSeriesAndEpisodes cmd = new CommandRequest_TvDBUpdateSeriesAndEpisodes(tvid, false);
-                cmd.Save();
-            }
-        }
-        */
+      
 
         public static void ScanForMatches()
         {
