+++ Shoko.Server/packages.config
@@ -8,7 +8,9 @@
   <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net462" />
   <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net462" />
   <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net462" />
+  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net462" />
   <package id="Microsoft.SQLServer.SMO" version="12.0.2000.8" targetFramework="net46" />
+  <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net462" />
   <package id="MySql.Data" version="6.9.9" targetFramework="net461" />
   <package id="Nancy" version="1.4.3" targetFramework="net462" />
   <package id="Nancy.Authentication.Stateless" version="1.4.1" targetFramework="net462" />
@@ -16,7 +18,8 @@
   <package id="Nancy.Rest.Annotations" version="1.4.3.1-beta" targetFramework="net462" />
   <package id="Nancy.Rest.Module" version="1.4.3.4-beta" targetFramework="net462" />
   <package id="Nancy.Serialization.JsonNet" version="1.4.1" targetFramework="net462" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net46" />
+  <package id="NETStandard.Library" version="1.6.1" targetFramework="net462" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net462" />
   <package id="NHibernate" version="4.1.1.4000" targetFramework="net462" />
   <package id="Nito.AsyncEx" version="4.0.1" targetFramework="net462" />
   <package id="NLog" version="4.4.3" targetFramework="net462" />
@@ -24,13 +27,51 @@
   <package id="protobuf-net" version="2.1.0" targetFramework="net461" />
   <package id="RestSharp" version="105.2.3" targetFramework="net461" />
   <package id="SharpZipLib" version="0.86.0" targetFramework="net461" />
+  <package id="System.AppContext" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
+  <package id="System.Console" version="4.3.0" targetFramework="net462" />
   <package id="System.Data.SQLite.Core" version="1.0.104.0" targetFramework="net461" />
+  <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net462" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.3.0" targetFramework="net461" />
+  <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="net462" />
+  <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="net462" />
+  <package id="System.Globalization" version="4.3.0" targetFramework="net462" />
+  <package id="System.Globalization.Calendars" version="4.3.0" targetFramework="net462" />
+  <package id="System.IO" version="4.3.0" targetFramework="net462" />
+  <package id="System.IO.Compression" version="4.3.0" targetFramework="net462" />
+  <package id="System.IO.Compression.ZipFile" version="4.3.0" targetFramework="net462" />
+  <package id="System.IO.FileSystem" version="4.3.0" targetFramework="net462" />
+  <package id="System.IO.FileSystem.Primitives" version="4.3.0" targetFramework="net462" />
+  <package id="System.Linq" version="4.3.0" targetFramework="net462" />
+  <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Net.Http" version="4.3.1" targetFramework="net462" />
+  <package id="System.Net.Primitives" version="4.3.0" targetFramework="net462" />
+  <package id="System.Net.Sockets" version="4.3.0" targetFramework="net462" />
+  <package id="System.ObjectModel" version="4.3.0" targetFramework="net462" />
+  <package id="System.Reflection" version="4.3.0" targetFramework="net462" />
+  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="net462" />
+  <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="net462" />
+  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.Handles" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net462" />
+  <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net462" />
   <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net461" />
   <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net461" />
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net461" />
   <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net461" />
+  <package id="System.Text.Encoding" version="4.3.0" targetFramework="net462" />
+  <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="net462" />
+  <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="net462" />
+  <package id="System.Threading" version="4.3.0" targetFramework="net462" />
+  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net462" />
+  <package id="System.Threading.Timer" version="4.3.0" targetFramework="net462" />
+  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net462" />
+  <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net462" />
   <package id="TaskScheduler" version="2.5.22" targetFramework="net462" />
   <package id="TMDbLib" version="1.0.0" targetFramework="net461" />
+  <package id="TvDbSharper" version="0.6.2" targetFramework="net462" />
 </packages>
